requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
backports-zoneinfo==0.2.1 ; python_version < "3.9"
    # via
    #   -r requirements.in
    #   django
build==0.10.0
    # via pip-tools
click==8.1.3
    # via pip-tools
colorama==0.4.6
    # via pytest-watch
coverage[toml]==7.2.3
    # via pytest-cov
django==4.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.14.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==23.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
docopt==0.6.2
    # via pytest-watch
exceptiongroup==1.1.1
    # via pytest
flake8==6.0.0
    # via
    #   -r requirements.in
    #   flake8-django
flake8-django==1.2
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
packaging==23.1
    # via
    #   build
    #   pytest
pip-tools==6.13.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2==2.9.6
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyjwt==1.7.1
    # via djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via build
pytest==7.3.0
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-watch
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-watch==4.2.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via
    #   build
    #   coverage
    #   pytest
watchdog==3.0.0
    # via pytest-watch
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
